requirements.txt
openai
pinecone
gradio
requests
python-dotenv
requests

PyPDF2
python-docx
pandas
openpyxl
ebooklib
beautifulsoup4
chardet
sentence-transformers
peft
transformers
torch
accelerate
bitsandbytes
langchain 
langgraph
python-redmine